+++ src/components/StudentsComponents/StudentRow/StudentRow.js
@@ -2,6 +2,7 @@ import React from 'react'
 import { CiEdit } from 'react-icons/ci'
 import { CiTrash } from 'react-icons/ci'
 import { useNavigate } from 'react-router-dom'
+import { confirmDialog } from 'primereact/confirmdialog'
 import './StudentRow.css'
 
 function StudentRow({ student, deleteStudent }) {
@@ -24,25 +25,35 @@ function StudentRow({ student, deleteStudent }) {
       <td className='student-row-item student-row-study-year'>
         {student.studyYear}
       </td>
-      <td>
-        <div className='student-row-item student-row-buttons'>
-          <div
-            className='student-row-edit'
-            onClick={(e) => {
-              navigate(`/students/manage-students/edit-student/${student.id}`)
-            }}
-          >
-            <CiEdit size={23} />
-          </div>
-          <div
-            className='student-row-delete'
-            onClick={() => {
-              deleteStudent(student.id)
-            }}
-          >
-            <CiTrash size={23} />
-          </div>
-        </div>
+      <td className='student-row-item student-row-buttons'>
+        <button
+          className='student-row-button student-row-edit'
+          onClick={(e) => {
+            navigate(`/students/manage-students/edit-student/${student.id}`)
+          }}
+        >
+          <CiEdit size={23} />
+        </button>
+        <button
+          className='student-row-button student-row-delete'
+          onClick={() => {
+            confirmDialog({
+              message: `Sunteți sigur că doriți să ștergeți studentul ${student.fullName}?`,
+              header: 'Confimare ștergere student',
+              icon: 'pi pi-info-circle',
+              defaultFocus: 'reject',
+              acceptClassName: 'p-button-danger',
+              acceptLabel: 'Da',
+              rejectLabel: 'Nu',
+              accept: () => {
+                deleteStudent(student.id)
+              },
+              reject: () => {},
+            })
+          }}
+        >
+          <CiTrash size={23} />
+        </button>
       </td>
     </tr>
   )
